SKFP-16: Fix review comment
